requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[argon2]
alembic
python-dotenv
pydantic-settings
pydantic[email]
slowapi